requirements.txt
coverage==4.5.1
coveralls==1.3.0
coreapi==2.3.3
dj-database-url==0.5.0
Django==1.11
django-extensions==2.0.6
django-filter==1.1.0
django-nose==1.4.5
django-rest-swagger==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==2.6.11
nose==1.3.7
psycopg2-binary==2.7.4
PyYAML==3.13
Werkzeug==0.14.1